% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ecoli_protein_classes}
\alias{ecoli_protein_classes}
\title{E. coli Protein Classes}
\format{
A data frame with X rows and 1 variable:
\describe{
  \item{class}{Integer value representing the protein classification: -1 for one class and 1 for the other}
}
}
\usage{
ecoli_protein_classes
}
\description{
A dataset containing protein classifications for E. coli, where each protein is categorized into one of two classes.
}
\keyword{datasets}
