% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ecoli_proteins}
\alias{ecoli_proteins}
\title{E. coli Protein Sequences}
\format{
An AAStringSet object of length 352:
\describe{
  \item{width}{The length of each protein sequence in amino acids}
  \item{seq}{The amino acid sequence of each protein}
  \item{names}{Unique identifiers for each protein sequence, in the format 511145.bXXXX}
}
}
\usage{
ecoli_proteins
}
\description{
A dataset containing amino acid sequences for 352 proteins from _Escherichia coli_.
}
\details{
The dataset includes the full amino acid sequences of proteins, each identified by a unique code.
Sequence information can be used for further bioinformatics analysis, such as alignment or functional prediction.
}
\keyword{datasets}
