% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{five_disease}
\alias{five_disease}
\title{Five Disease Data}
\format{
A tibble with 2,406 rows and 5 variables:
\describe{
  \item{disease}{Character string representing the disease (e.g., "Alzheimers")}
  \item{a}{Numeric value representing a measurement related to the disease}
  \item{b}{Numeric value representing another measurement related to the disease}
  \item{c}{Numeric value representing a third measurement related to the disease}
  \item{gene}{Character string representing the gene associated with the observation}
}
}
\usage{
five_disease
}
\description{
A dataset containing gene expression or related data across five diseases.
}
\details{
This dataset can be used for analyzing variations in gene measurements across different diseases.
}
\keyword{datasets}
