% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hglobin}
\alias{hglobin}
\title{Hemoglobin Protein Sequences}
\format{
An AAStringSet object of length 3:
\describe{
  \item{width}{The length of each hemoglobin protein sequence in amino acids}
  \item{seq}{The amino acid sequence of each hemoglobin protein}
  \item{names}{Species identifiers for each protein sequence: "HBA_HUMAN" for human, "HBA_MOUSE" for mouse, and "HBAZ_CAPHI" for Capra hircus (goat)}
}
}
\usage{
hglobin
}
\description{
A data set containing amino acid sequences for hemoglobin proteins from different species.
}
\details{
The dataset includes the full amino acid sequences for hemoglobin proteins from human, mouse, and goat. These sequences can be used for comparative analysis or structural studies.
}
\keyword{datasets}
