% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Example Metadata for Genomic Data}
\format{
A data frame with 1,829 rows and 22 variables:
\describe{
  \item{DepMap_ID}{Character string representing the unique identifier for each cell line.}
  \item{cell_line_name}{Character string representing the name of the cell line.}
  \item{stripped_cell_line_name}{Character string representing a simplified version of the cell line name.}
  \item{CCLE_Name}{Character string representing the CCLE name.}
  \item{alias}{Character string for alternative names for the cell line.}
  \item{COSMICID}{Integer representing the COSMIC ID.}
  \item{sex}{Character string indicating the sex of the cell line ("Male" or "Female").}
  \item{source}{Character string representing the source of the cell line (e.g., "ATCC").}
  \item{RRID}{Character string for the Research Resource Identifier.}
  \item{WTSI_Master_Cell_ID}{Integer representing the WTSI master cell ID.}
  \item{sample_collection_site}{Character string indicating the sample collection site.}
  \item{primary_or_metastasis}{Character string indicating if the sample is primary or metastatic.}
  \item{primary_disease}{Character string for the primary disease associated with the cell line.}
  \item{Subtype}{Character string for the subtype of the disease.}
  \item{age}{Numeric value indicating the age of the donor, if available.}
  \item{Sanger_Model_ID}{Character string for the Sanger model ID.}
  \item{depmap_public_comments}{Character string for any public comments about the cell line.}
  \item{lineage}{Character string indicating the lineage (e.g., "blood").}
  \item{lineage_subtype}{Character string for the lineage subtype.}
  \item{lineage_sub_subtype}{Character string for the sub-subtype of the lineage.}
  \item{lineage_molecular_subtype}{Character string for the molecular subtype within the lineage.}
  \item{culture_type}{Character string indicating the culture type (e.g., "Adherent", "Suspension").}
}
}
\usage{
metadata
}
\description{
A dataset containing example metadata for genomic data analysis, including details about cell lines and related attributes.
}
\details{
This dataset is used to illustrate how metadata can be used in genomic analysis, including characteristics of cell lines and their related information.
}
\keyword{datasets}
