% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{plastid_genomes}
\alias{plastid_genomes}
\title{Plastid Genomes Dataset}
\format{
A `DNAStringSet` object with the following structure:
\describe{
  \item{pool}{Contains raw sequence data stored as a `SharedRaw_Pool` object}
  \item{ranges}{Contains information on each sequence, including:
    \describe{
      \item{NAMES}{Names of each sequence, including accession numbers and species (e.g., "NC_018523.1 Saccharina japonica chloroplast, complete genome")}
    }
  }
  \item{elementType}{Indicates the element type, "DNAString" for each sequence}
}
}
\usage{
plastid_genomes
}
\description{
A `DNAStringSet` object containing complete chloroplast genome sequences from different plant species.
}
\details{
The dataset includes chloroplast genome sequences for five species, each represented by a complete genome sequence in DNA format. Useful for comparative genomics and evolutionary studies.
}
\keyword{datasets}
