% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ts_data}
\alias{ts_data}
\title{Time Series SNP Data}
\format{
A data frame with 400 rows and 3 variables:
\describe{
  \item{day}{Integer indicating the day of the observation}
  \item{snp}{Character string representing the SNP identifier}
  \item{frequency}{Numeric value representing the frequency of the SNP on the given day}
}
}
\usage{
ts_data
}
\description{
A dataset containing time series data for SNP frequencies over 400 days.
}
\details{
This dataset is used to analyze the changes in SNP frequencies over time.
}
\keyword{datasets}
